Use dpkg-source --skip-patches
Instead of trying to unapply patches later, ask dpkg-source to not apply
them in the first place.

This works for "3.0 (quilt)" format packages. For 1.0 packages, a
warning is printed.

It seems a bit tricky to detect the format before extraction. grep-dctrl
on the .dsc file works fine, but only if the .dsc file is not signed.